fix(api): move parentGroupAuditAttrs to server_groups.go
The helper is only called within server_groups.go; keeping it in server.go
creates an implicit dependency that undermines the purpose of the file split.
Moving it next to its callers makes the dependency explicit and keeps
server.go free of group-specific logic.
